backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
llama-index>=0.8.0
pycaret>=3.0.0
pandas>=2.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0